refactor: pull out to file
