+++ include/pybind11/gil.h
@@ -199,43 +199,3 @@ private:
 PYBIND11_NAMESPACE_END(PYBIND11_NAMESPACE)
 
 #endif // !PYBIND11_SIMPLE_GIL_MANAGEMENT
-
-PYBIND11_NAMESPACE_BEGIN(PYBIND11_NAMESPACE)
-
-class scoped_critical_section {
-public:
-    /// This does not do anything if there's a GIL. On free-threaded Python,
-    /// it locks an object. This uses the CPython API, which has limits
-    scoped_critical_section(handle obj) : single(true) {
-#ifdef Py_GIL_DISABLED
-        PyCriticalSection_Begin(&section, obj.ptr());
-#endif
-    }
-
-    scoped_critical_section(handle obj1, handle obj2) : single(true) {
-#ifdef Py_GIL_DISABLED
-        PyCriticalSection2_Begin(&section2, obj1.ptr(), obj2.ptr());
-#endif
-    }
-
-    ~scoped_critical_section() {
-#ifdef Py_GIL_DISABLED
-        if(single) {
-            PyCriticalSection_End(&section);
-        } else {
-            PyCriticalSection2_End(&section2);
-        }
-#endif
-    }
-
-private:
-    bool single;
-#ifdef Py_GIL_DISABLED
-    union {
-        PyCriticalSection section;
-        PyCriticalSection2 section2;
-    };
-#endif
-};
-
-PYBIND11_NAMESPACE_END(PYBIND11_NAMESPACE)
