+++ tests/extra_python_package/test_files.py
@@ -44,6 +44,7 @@ main_headers = {
     "include/pybind11/chrono.h",
     "include/pybind11/common.h",
     "include/pybind11/complex.h",
+    "include/pybind11/critical_section.h",
     "include/pybind11/eigen.h",
     "include/pybind11/embed.h",
     "include/pybind11/eval.h",
