+++ tests/test_embed/test_interpreter.cpp
@@ -1,4 +1,5 @@
 #include <pybind11/embed.h>
+#include <pybind11/critical_section.h>
 
 // Silence MSVC C++17 deprecation warning from Catch regarding std::uncaught_exceptions (up to
 // catch 2.0.1; this should be fixed in the next catch release after 2.0.1).
